chore: re-organize source code directory hierarchy
