+++ test/form_widdget_test.dart
@@ -1,10 +1,10 @@
-import 'package:cabinet/components/widgets/form_field_item.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:cabinet/components/widgets/form_widget.dart';
-import 'package:form_builder_validators/form_builder_validators.dart';
+import 'package:cabinet/widgets/form_field_item.dart';
+import 'package:cabinet/widgets/form_widget.dart';
 
 void main() {
   testWidgets('should render empty form widget if no fields are provided', (WidgetTester tester) async {
